Add a script to measure performance against Mocha and Jest
