+++ performance.sh
@@ -1,6 +1,8 @@
 #! /bin/bash
 
-npx mocha -t 100000 ./test/performance/mocha.js
+npx mocha -t 100000 --reporter min ./test/performance/mocha.js
+
+jest --testPathPattern=performance --reporters jest-silent-reporter
 
 npx titef ./test/performance/titef.js
